chore(suite-desktop): separate getProcessIcon for future reuse
